fix: scan status, also in compare
