+++ app/components/analysis-details.js
@@ -35,10 +35,6 @@ const AnalysisDetailsComponent = Ember.Component.extend({
     return filteredRisks[0].value;
   }).property("filteredRisks"),
 
-  tScanning: t("scanning"),
-  tUntested: t("untested"),
-  tRequested: t("requested"),
-
   riskClass: ( function() {
     const risk = this.get("analysis.computedRisk");
     switch (risk) {
@@ -111,49 +107,6 @@ const AnalysisDetailsComponent = Ember.Component.extend({
     "analysis.file.isDynamicDone",
     "analysis.file.isManualDone",
     "analysis.file.isApiDone"
-  ),
-
-  scanningText: (function() {
-    const tScanning = this.get("tScanning");
-    const tUntested = this.get("tUntested");
-    const tRequested = this.get("tRequested");
-    const types = this.get("analysis.vulnerability.types");
-    if (types === undefined) { return []; }
-    switch (types[0]) {
-      case ENUMS.VULNERABILITY_TYPE.STATIC:
-        return tScanning;
-      case ENUMS.VULNERABILITY_TYPE.DYNAMIC:
-        const dynamicStatus = this.get('analysis.file.dynamicStatus');
-        if(dynamicStatus !== ENUMS.DYNAMIC_STATUS.NONE) {
-          return tScanning;
-        }
-        else {
-          return tUntested;
-        }
-        break;
-      case ENUMS.VULNERABILITY_TYPE.MANUAL:
-        if(this.get("analysis.file.manual")) {
-          return tRequested;
-        }
-        else {
-          return tUntested;
-        }
-        break;
-      case ENUMS.VULNERABILITY_TYPE.API:
-        const apiScanProgress = this.get('analysis.file.apiScanProgress');
-        if(apiScanProgress >= 1) {
-          return tScanning;
-        }
-        else {
-          return tUntested;
-        }
-        break;
-    }
-  }).property(
-    "analysis.vulnerability.types",
-    "analysis.file.dynamicStatus",
-    "analysis.file.manual",
-    "analysis.file.apiScanProgress"
   ),
 
   actions: {
