+++ app/models/analysis.js
@@ -1,6 +1,7 @@
+import Ember from 'ember';
 import DS from 'ember-data';
 import ENUMS from 'irene/enums';
-import Ember from 'ember';
+import { translationMacro as t } from 'ember-i18n';
 
 const Analysis = DS.Model.extend({
   findings: DS.attr(),
