+++ app/styles/app.sass
@@ -125,6 +125,9 @@ body
 .margin-side-auto
   margin: 0px auto
 
+.width-100
+  width: 100px
+
 /* Appknox specific */
 .appknox-logo
   margin-top: 13px
