+++ app/templates/components/scanning-text.emblem
@@ -1,0 +1,1 @@
+| &nbsp;{{scanningText}}
