+++ tests/integration/components/scanning-text-test.js
@@ -1,0 +1,25 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+moduleForComponent('scanning-text', 'Integration | Component | scanning text', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });
+
+  this.render(hbs`{{scanning-text}}`);
+
+  assert.equal(this.$().text().trim(), '');
+
+  // Template block usage:
+  this.render(hbs`
+    {{#scanning-text}}
+      template block text
+    {{/scanning-text}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+});
